Pickled data and code to read-in
After abusing my RAM a few times, I looked for more efficient ways of reading in the data. I came across "pickling", which is basically converting the pandas dataframe into a bitstream. This bitstream was then compressed using the gzip algorithm, so each day is only about 5-6MB. 

I've included the code I used to produce the pickled files as well as code to read it in. I've noticed it is slightly faster to read in and uses slightly less memory using this method.